requirements.txt
amqp==2.1.4
astroid==1.5.3
billiard==3.5.0.2
celery==4.0.2
decorator==4.0.11
Django==1.11.2
django-extensions==1.7.9
django-filter==1.0.4
django-model-utils==3.0.0
djangorestframework==3.6.3
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.10.2
kombu==4.0.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pylint==1.7.1
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytz==2017.2
redis==2.10.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
vine==1.1.3
wcwidth==0.1.7
wrapt==1.10.10
